readme
Pitch shifting is another process to artificially increase the audio data. For doing pitch shifting, audio data is first time stretch based on time stretching algorithm from figure (11). The time stretch audio is then resampled based on target sampling rate. The length of the audio is decreases if the target sampling rate is less than original sampling rate whereas the length is increases if the target sampling rate is more than original sampling rate. The increased length of the audio is cropped so that the audio resampled by target sampling rate is equal to the original audio. The algorithm for pitch shifting is explain in Fig. 21.